+++ packages/ui-kit/stories/components/inputs/checkbox/checkbox.stories.tsx
@@ -1,17 +1,21 @@
 import { useState } from 'react'
 
-import { Meta } from '@storybook/react'
-
 import { Checkbox } from '../../../../src'
 import { VerticalContainer } from '../../../../storybook-utils/components/containers/vertical'
 import { ValuePreview } from '../../../../storybook-utils/components/previews/value-preview'
+import { Meta } from '@storybook/react'
 
 export default {
-  title: 'Components/Data Entry/Checkbox',
   component: Checkbox,
+  title: 'Components/Data Entry/Checkbox',
 } as Meta<typeof Checkbox>
 
 export const Default = {
+  args: {
+    disabled: false,
+    label: 'Click here',
+  },
+
   render: args => {
     const [checked, setChecked] = useState(true)
 
@@ -22,9 +26,30 @@ export const Default = {
       </VerticalContainer>
     )
   },
-
+}
+export const Tight = {
   args: {
-    label: 'Click here',
     disabled: false,
+    tight: true,
+  },
+
+  render: args => {
+    return (
+      <>
+        <Checkbox
+          {...args}
+          label={
+            'Указано направление, в котором хотел бы развиваться (более глубокое изучение React/изучение других front-end фреймворков/full-stack разработка/разработка для мобильных устройств/что-то другое). Указывай только одно направление желаемого развития'
+          }
+        />
+        <Checkbox
+          {...args}
+          label={
+            'Указано направление, в котором хотел бы развиваться (более глубокое изучение React/изучение других front-end фреймворков/full-stack разработка/разработка для мобильных устройств/что-то другое). Указывай только одно направление желаемого развития'
+          }
+        />
+        <Checkbox {...args} label={'Указано направление,'} />
+      </>
+    )
   },
 }